% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOutCell.R
\name{checkOutCell}
\alias{checkOutCell}
\title{Check out a cell.}
\usage{
checkOutCell(userID, tblDir = "//lar-file-srv/Data/BTPD_2016/Digitizing",
  shpDir = "//lar-file-srv/Data/BTPD_2016/Analysis/data/Shapefiles/BTPD_Grid_CO_Ranked",
  shp = "BTPD_Grid_CO_Ranked")
}

