\name{pDog-package}
\alias{pDog-package}
\alias{pDog}

\docType{package}

\title{Manage prairie-dog project.}

\description{
pDog is a package.
}

\details{
\tabular{ll}{
Package: \tab pDog\cr
Type: \tab Package\cr
License: \tab GNU General Public License\cr
Dependencies: \tab maptools, raster, rgdal, rgeos\cr 
}
This is a package.

}

\author{
Jason Mitchell \code{jmitchell@west-inc.com}.
}

\keyword{ package }
